Move CoreUpdater UI tests to plugin
