+++ plugins/CoreUpdater/tests/Fixtures/DbUpdaterTestFixture.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\CoreUpdater\Test\Fixtures;
+namespace Piwik\Plugins\CoreUpdater\tests\Fixtures;
 
 use Piwik\Tests\Fixtures\SqlDump;
