+++ plugins/CoreUpdater/tests/Integration/ReleaseChannelTest.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\CoreUpdater\Test\ReleaseChannel;
+namespace Piwik\Plugins\CoreUpdater\tests\ReleaseChannel;
 
 use Piwik\Config;
 use Piwik\Db;
