+++ plugins/CoreUpdater/tests/Integration/UpdateCommunicationTest.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\CoreUpdater\Test;
+namespace Piwik\Plugins\CoreUpdater\tests\Integration;
 
 use Piwik\Config;
 use Piwik\Option;
