+++ plugins/CoreUpdater/tests/Mock/UpdaterMock.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\CoreUpdater\Test\Mock;
+namespace Piwik\Plugins\CoreUpdater\tests\Mock;
 
 use Piwik\Plugins\CoreUpdater\ArchiveDownloadException;
 use Piwik\Plugins\CoreUpdater\Updater;
