+++ plugins/CoreUpdater/tests/UI/expected-screenshots/CoreUpdaterCode_newVersion.png
