+++ plugins/CoreUpdater/tests/UI/expected-screenshots/CoreUpdaterDb_main.png
