+++ plugins/CoreUpdater/tests/UI/expected-screenshots/CoreUpdaterDb_main_instance.png
