+++ plugins/CoreUpdater/tests/UI/expected-screenshots/CoreUpdaterDb_updated.png
