+++ plugins/CoreUpdater/tests/Unit/ModelTest.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\CoreUpdater\Test\Unit;
+namespace Piwik\Plugins\CoreUpdater\tests\Unit;
 
 use Piwik\Plugins\CoreUpdater\Model;
