+++ plugins/MultiSites/tests/UI/MultiSites_spec.js
@@ -1,7 +1,7 @@
 /*!
  * Matomo - free/libre analytics platform
  *
- * Screenshot integration tests ofr MultiSites.
+ * Screenshot integration tests for MultiSites.
  *
  * @link https://matomo.org
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
